Add team, class selector, qr

+++ app/Http/Controllers/Auth/UserController.php
@@ -6,12 +6,17 @@ use App\Http\Controllers\{
     Controller
 };
 use App\Models\{
-    User
+    User,
+    EJGClass
 };
+
+use Illuminate\Http\Request;
+
 
 use Illuminate\Support\Facades\{
     Auth,
-    Redirect
+    Redirect,
+    Gate
 };
 
 
@@ -19,24 +24,43 @@ class UserController extends Controller
 {
 
     public function index(){
+        return redirect()->route('index');
         return view('user.index');
     }
 
-    public function edit(){
-        /**
-         * @var User
-         */
-        $user = Auth::login();
+    public function edit(Request $request, $userID){
+        $user = User::findOrFail($userID);
+        Gate::authorize('update',$user);
 
-        $user->save();
+        return view('user.edit',["user"=>$user,"classes"=>EJGClass::all()]);
     }
 
-    public function destroy(){
-        $user = User::find(Auth::user()->id);
-        Auth::logout();
-        if ($user->delete()) {
-            return Redirect::route('home');
-       }
+    public function update(Request $request, $userID){
+        $user = User::findOrFail($userID);
+        Gate::authorize('update',$user);
+
+        if($user->ejgClass==null){
+            $class = EJGClass::findOrFail($request->input('ejgClass'));
+            $user->ejgClass()->associate($class);
+            $user->save();
+        }
+
+        return redirect()->route('index');
+    }
+
+    public function destroy(Request $request, $userID){
+        $user = User::findOrFail($userID);
+        Gate::authorize('delete',$user);
+        if($user->is($request->user())){
+            Auth::logout();
+            if ($user->delete()) {
+                return redirect()->route('index');
+           }
+        }
+        else{
+            $user->delete();
+            return back();
+        }
 
     }
 }
